refactor: move sync logic to new comp (#605)
